+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/tasks/CallbackTiming.java
@@ -4,30 +4,30 @@ import org.ovirt.engine.core.bll.tasks.interfaces.CommandCallback;
 
 class CallbackTiming {
     // Total delay between callback executions
-    private int initialDelay;
+    private long initialDelay;
 
     // Remaining delay to next callback execution
-    private int remainingDelay;
+    private long remainingDelay;
     private CommandCallback callback;
 
     // the end-time, where the callback shouldn't wait for the event any longer and change to polling mode
     private long waitOnEventEndTime;
 
-    public CallbackTiming(CommandCallback callback, int executionDelay) {
+    public CallbackTiming(CommandCallback callback, long executionDelay) {
         this.callback = callback;
         this.initialDelay = executionDelay;
         this.remainingDelay = executionDelay;
     }
 
-    public int getInitialDelay() {
+    public long getInitialDelay() {
         return initialDelay;
     }
 
-    public void setInitialDelay(int initialDelay) {
+    public void setInitialDelay(long initialDelay) {
         this.initialDelay = initialDelay;
     }
 
-    public int getRemainingDelay() {
+    public long getRemainingDelay() {
         return remainingDelay;
     }
 
@@ -39,7 +39,7 @@ class CallbackTiming {
         return waitOnEventEndTime;
     }
 
-    public void setRemainingDelay(int remainingDelay) {
+    public void setRemainingDelay(long remainingDelay) {
         this.remainingDelay = remainingDelay;
     }
